+++ src/JsonnetNativeCallback.cpp
@@ -8,14 +8,14 @@ namespace nodejsonnet {
 
     NativeCallbackPayload::NativeCallbackPayload(
       std::shared_ptr<JsonnetVm> vm, std::vector<JsonnetJsonValue const *> args)
-      : CallbackPayload{std::move(vm)}, args{std::move(args)} {}
-
+      : CallbackPayload{std::move(vm)}, args{std::move(args)} {
+    }
 
     std::vector<napi_value> NativeCallbackPayload::makeArgs(Napi::Env env) const {
       JsonValueConverter const conv{getVm()};
       std::vector<napi_value> ret;
       ret.reserve(args.size());
-      for(auto const arg : args) {
+      for(auto const arg: args) {
         ret.push_back(conv.toNapiValue(env, arg));
       }
       return ret;